+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display handle with '@' in account link tooltips.
 
+### Fixed
+
+- Don't allow profile image upload if media type is not supported.
+
 ## [2.13.0] - 2024-03-12
 
 ### Added